% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{check_domain_transferability}
\alias{check_domain_transferability}
\title{Checks whether a domain name can be transferred to Amazon Route 53}
\usage{
check_domain_transferability(DomainName, AuthCode = NULL)
}
\arguments{
\item{DomainName}{The name of the domain that you want to transfer to Amazon Route 53.

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.}

\item{AuthCode}{If the registrar for the top-level domain (TLD) requires an authorization code to transfer the domain, the code that you got from the current registrar for the domain.}
}
\description{
Checks whether a domain name can be transferred to Amazon Route 53.
}
\examples{

}
