% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{delete_tags_for_domain}
\alias{delete_tags_for_domain}
\title{This operation deletes the specified tags for a domain}
\usage{
delete_tags_for_domain(DomainName, TagsToDelete)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to delete one or more tags.}

\item{TagsToDelete}{[required] A list of tag keys to delete.}
}
\description{
This operation deletes the specified tags for a domain.
}
\details{
All tag operations are eventually consistent; subsequent operations might not immediately represent all issued operations.
}
\section{Accepted Parameters}{
\preformatted{delete_tags_for_domain(
  DomainName = "string",
  TagsToDelete = list(
    "string"
  )
)
}
}

