% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{enable_domain_transfer_lock}
\alias{enable_domain_transfer_lock}
\title{This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status) to prevent domain transfers}
\usage{
enable_domain_transfer_lock(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to set the transfer lock for.}
}
\description{
This operation sets the transfer lock on the domain (specifically the \code{clientTransferProhibited} status) to prevent domain transfers. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.
}
\section{Accepted Parameters}{
\preformatted{enable_domain_transfer_lock(
  DomainName = "string"
)
}
}

