% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{get_contact_reachability_status}
\alias{get_contact_reachability_status}
\title{For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation returns information about whether the registrant contact has responded}
\usage{
get_contact_reachability_status(domainName = NULL)
}
\arguments{
\item{domainName}{The name of the domain for which you want to know whether the registrant contact has confirmed that the email address is valid.}
}
\description{
For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation returns information about whether the registrant contact has responded.
}
\details{
If you want us to resend the email, use the \code{ResendContactReachabilityEmail} operation.
}
\section{Accepted Parameters}{
\preformatted{get_contact_reachability_status(
  domainName = "string"
)
}
}

