% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{get_domain_detail}
\alias{get_domain_detail}
\title{This operation returns detailed information about a specified domain that is associated with the current AWS account}
\usage{
get_domain_detail(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to get detailed information about.}
}
\description{
This operation returns detailed information about a specified domain that is associated with the current AWS account. Contact information for the domain is also returned as part of the output.
}
\section{Accepted Parameters}{
\preformatted{get_domain_detail(
  DomainName = "string"
)
}
}

