% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{list_domains}
\alias{list_domains}
\title{This operation returns all the domain names registered with Amazon Route 53 for the current AWS account}
\usage{
list_domains(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for \code{MaxItems}, you can use \code{Marker} to return additional domains. Get the value of \code{NextPageMarker} from the previous response, and submit another request that includes the value of \code{NextPageMarker} in the \code{Marker} element.

Constraints: The marker must match the value specified in the previous request.}

\item{MaxItems}{Number of domains to be returned.

Default: 20}
}
\description{
This operation returns all the domain names registered with Amazon Route 53 for the current AWS account.
}
\examples{

}
