% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{list_tags_for_domain}
\alias{list_tags_for_domain}
\title{This operation returns all of the tags that are associated with the specified domain}
\usage{
list_tags_for_domain(
  DomainName = "string"
)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to get a list of tags.}
}
\description{
This operation returns all of the tags that are associated with the specified domain.
}
\details{
All tag operations are eventually consistent; subsequent operations might not immediately represent all issued operations.
}
\examples{

}
