% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_package.R
\docType{package}
\name{paws.route53domains-package}
\alias{paws.route53domains}
\alias{paws.route53domains-package}
\title{paws.route53domains: Amazon Route 53 Domains}
\description{
Amazon Route 53 API actions let you register domain
    names and perform related operations.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
