% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{register_domain}
\alias{register_domain}
\title{This operation registers a domain}
\usage{
register_domain(DomainName, DurationInYears, AdminContact, RegistrantContact,
  TechContact, IdnLangCode = NULL, AutoRenew = NULL,
  PrivacyProtectAdminContact = NULL, PrivacyProtectRegistrantContact = NULL,
  PrivacyProtectTechContact = NULL)
}
\arguments{
\item{DomainName}{The domain name that you want to register.

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.}

\item{DurationInYears}{The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see \href{http://docs.aws.amazon.com/Route53/latest/DeveloperGuide/registrar-tld-list.html}{Domains that You Can Register with Amazon Route 53} in the \emph{Amazon Route 53 Developer Guide}.

Default: 1}

\item{AdminContact}{Provides detailed contact information.}

\item{RegistrantContact}{Provides detailed contact information.}

\item{TechContact}{Provides detailed contact information.}

\item{IdnLangCode}{Reserved for future use.}

\item{AutoRenew}{Indicates whether the domain will be automatically renewed (\code{true}) or not (\code{false}). Autorenewal only takes effect after the account is charged.

Default: \code{true}}

\item{PrivacyProtectAdminContact}{Whether you want to conceal contact information from WHOIS queries. If you specify \code{true}, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify \code{false}, WHOIS queries return the information that you entered for the admin contact.

Default: \code{true}}

\item{PrivacyProtectRegistrantContact}{Whether you want to conceal contact information from WHOIS queries. If you specify \code{true}, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify \code{false}, WHOIS queries return the information that you entered for the registrant contact (the domain owner).

Default: \code{true}}

\item{PrivacyProtectTechContact}{Whether you want to conceal contact information from WHOIS queries. If you specify \code{true}, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify \code{false}, WHOIS queries return the information that you entered for the technical contact.

Default: \code{true}}
}
\description{
This operation registers a domain. Domains are registered either by Amazon Registrar (for .com, .net, and .org domains) or by our registrar associate, Gandi (for all other domains). For some top-level domains (TLDs), this operation requires extra parameters.
}
\details{
When you register a domain, Amazon Route 53 does the following:
\itemize{
\item Creates a Amazon Route 53 hosted zone that has the same name as the domain. Amazon Route 53 assigns four name servers to your hosted zone and automatically updates your domain registration with the names of these name servers.
\item Enables autorenew, so your domain registration will renew automatically each year. We'll notify you in advance of the renewal date so you can choose whether to renew the registration.
\item Optionally enables privacy protection, so WHOIS queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you don't enable privacy protection, WHOIS queries return the information that you entered for the registrant, admin, and tech contacts.
\item If registration is successful, returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant is notified by email.
\item Charges your AWS account an amount based on the top-level domain. For more information, see \href{http://aws.amazon.com/route53/pricing/}{Amazon Route 53 Pricing}.
}
}
\examples{

}
