% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{retrieve_domain_auth_code}
\alias{retrieve_domain_auth_code}
\title{This operation returns the AuthCode for the domain}
\usage{
retrieve_domain_auth_code(DomainName)
}
\arguments{
\item{DomainName}{The name of the domain that you want to get an authorization code for.}
}
\description{
This operation returns the AuthCode for the domain. To transfer a domain to another registrar, you provide this value to the new registrar.
}
\examples{

}
