% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{update_domain_nameservers}
\alias{update_domain_nameservers}
\title{This operation replaces the current set of name servers for the domain with the specified set of name servers}
\usage{
update_domain_nameservers(
  DomainName = "string",
  FIAuthKey = "string",
  Nameservers = list(
    list(
      Name = "string",
      GlueIps = list(
        "string"
      )
    )
  )
)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to change name servers for.}

\item{FIAuthKey}{The authorization key for .fi domains}

\item{Nameservers}{[required] A list of new name servers for the domain.}
}
\description{
This operation replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.
}
\details{
If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.
}
\examples{

}
