% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53domains_operations.R
\name{update_tags_for_domain}
\alias{update_tags_for_domain}
\title{This operation adds or updates tags for a specified domain}
\usage{
update_tags_for_domain(DomainName, TagsToUpdate = NULL)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to add or update tags.}

\item{TagsToUpdate}{A list of the tag keys and values that you want to add or update. If you specify a key that already exists, the corresponding value will be replaced.}
}
\description{
This operation adds or updates tags for a specified domain.
}
\details{
All tag operations are eventually consistent; subsequent operations might not immediately represent all issued operations.
}
\section{Accepted Parameters}{
\preformatted{update_tags_for_domain(
  DomainName = "string",
  TagsToUpdate = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

